requirements.txt
accelerate==0.25.0
diffusers==0.20.0
einops==0.6.1
lightning-utilities==0.9.0
matplotlib==3.7.3
nltk==3.8.1
numpy
opencv-python==4.8.1.78
Pillow==9.4.0
pytorch-lightning==2.0.7
scikit-image==0.22.0
scikit-learn==1.3.1
torch==2.2.0
torch-kmeans==0.2.0
torchvision==0.17.0
transformers==4.32.1
spaces==0.24.2
